backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
psycopg2-binary==2.9.11
python-dotenv==1.0.1
requests==2.31.0
Flask-SocketIO==5.3.6
openai==1.14.2
